% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{useapitest_api_get}
\alias{useapitest_api_get}
\title{GET Use API Test API}
\usage{
useapitest_api_get(...)
}
\arguments{
\item{...}{Path and query string components (endpoint parameters) should be
supplied here. If unnamed, the first object is assumed to be the API path
path, which is the string pointing to the desired API endpoint
\code{users/self}. Additional named arguments supplied here will be included
as part of the query string (trailing the "?" in the URL).}
}
\value{
An HTTP response object.
}
\description{
Send GET requests to Use API Test's API
}
\examples{
\dontrun{
## make custom request to locations/search endpoint
useapitest_locs <- useapitest_api_get("path/name", arg1 = TRUE, arg2 = 0)

## view data
useapitest_as_tbl(useapitest_locs)
}
}
