% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{useapitest_as_tbl}
\alias{useapitest_as_tbl}
\title{Convert useapitest output to tibble}
\usage{
useapitest_as_tbl(x)
}
\arguments{
\item{x}{Output from useapitest function}
}
\value{
Print out of summary info and a tibble of the data.
}
\description{
Convert useapitest response objects into informative tibbles
}
