% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{useapitest_token}
\alias{useapitest_token}
\alias{useapitest_create_token}
\title{useapitest_token}
\usage{
useapitest_token()

useapitest_create_token(access_token = NULL, client_id = NULL,
  client_secret = NULL, ...)
}
\arguments{
\item{access_token}{Access token. Token creation should occur via supplying
access token directly or via client ID and client secret. You should
provide one or the other to this function.}

\item{client_id}{Client key. Token creation should occur via supplying
access token directly or via client ID and client secret. You should
provide one or the other to this function.}

\item{client_secret}{Client secret. Required if access token is NULL and
a value is supplied for client_id.}

\item{...}{Other args passed to \link[httr]{init_oauth2.0}}
}
\value{
The user's access token, if available.

Sets environment variable and invisibly returns access token.
}
\description{
Accessing the user's instagram token

Creating access token for interacting with Instagram API
}
